Dockerfile update and package restructuring
