+++ ingest/Dockerfile
@@ -13,32 +13,32 @@ RUN apt-get update \
 	&& apt-get autoremove -y \
 	&& apt-get clean \
 	&& rm -rf /var/lib/apt/lists/*
-
 COPY "./protobuf/datastore.proto" "/protobuf/datastore.proto"
 COPY "./api" "${DOCKER_PATH}/api/"
-COPY "requirements.txt" "${DOCKER_PATH}/"
-COPY "./src/" "/ingest/src"
-COPY "./pyproject.toml" "/ingest/pyproject.toml"
-COPY "./setup.py" "/ingest/"
+COPY "requirements.txt" "${DOCKER_PATH}/api/"
+COPY "./src/" "${DOCKER_PATH}/src/"
+COPY "./pyproject.toml" "${DOCKER_PATH}/pyproject.toml"
+COPY "./setup.py" "${DOCKER_PATH}/setup.py"
 
-RUN pip install --no-cache-dir --upgrade -r "${DOCKER_PATH}/requirements.txt"
+RUN pip install --no-cache-dir --upgrade -r "${DOCKER_PATH}/api/requirements.txt"
 RUN pip install --no-cache-dir --upgrade pybind11
 
 # hadolint ignore=DL3013
-WORKDIR /ingest/src/ingest/bufr/
+WORKDIR "${DOCKER_PATH}"/src/ingest/bufr/
 RUN make
 WORKDIR /
 
-# hadolint ignore=DL3013
-RUN pip install --no-cache-dir --upgrade pip \
-	&& pip install --no-cache-dir --upgrade "/ingest/"
 # Compiling the protobuf file
 RUN python -m grpc_tools.protoc  \
 	--proto_path="protobuf" "protobuf/datastore.proto" \
 	--python_out="${DOCKER_PATH}"  \
 	--grpc_python_out="${DOCKER_PATH}"
 
-
-
 WORKDIR "${DOCKER_PATH}"
+
+RUN python "${DOCKER_PATH}/api/generate_standard_name.py"
+
+# hadolint ignore=DL3013
+RUN pip install --no-cache-dir --upgrade pip \
+	&& pip install --no-cache-dir --upgrade "${DOCKER_PATH}/"
 CMD ["gunicorn", "api.main:app", "--workers=4", "--worker-class=uvicorn.workers.UvicornWorker", "--bind=0.0.0.0:8001"]
