+++ ingest/requirements.txt
@@ -8,3 +8,4 @@ fastapi == 0.108.0
 grpcio-tools==1.58.0
 uvicorn==0.25.0
 python-multipart~=0.0.9
+requests == 2.31.0
